Add files via upload
Online teaching is a rather new concept which has recently come to day-to-day practice making the whole process not so friendly for the teachers who have been usings chalks and markers and students who are habitual of studying in a real time classroom. Writing on an online board is a task that mostly no teacher/ user of online teaching platforms can avoid. It is generally rather difficult for users to write using a standard mouse/ mousepad. The uneasiness of the whole process consumes a lot of time and makes the classroom less efficient by minimising the output. Stylist does make the job easier but it increases the input cost and adds on to the cluttering of the workspace also taking extra space in the laptop bag. Similarly, there are various challenges that users of online teaching platforms all over the world face which are not been dealt with eagerness. 
 
The process of online teaching, office meetings, presentations, etc will become a lot easier if the minimum input required would just be the user and some room light. The proposed project aims in minimising the hassles listed above by introducing gesture and body recognition as a major method of providing commands to the computer. 
The common problem of writing on online boards faced by users is solved by simply providing the option of writing directly on the screen by using the user’s mere finger as a pointer and palm as an eraser. This makes the whole teaching experience more time sufficient, user friendly, interactive and productive. 
The gesture recognition method also makes multitasking easier. Eg- a user can drag an image on the screen, while resizing it accordingly and can also write or erase on the screen/ image at the same time. This process is not limited to images and can be used simultaneously with pdf files, slideshows, word files, etc. 
 
This method removes the responsibility from the user to own the appropriate input units and use them correctly to give commands. By using the gesture recognition methods, the user can give commands in time to the machine and doesn’t really have to worry about learning to use the various input units and wait for them to ask the machine to perform a certain task.